feat: modify modules file and index.js
