+++ routes/pages/modules/users.js
@@ -1,5 +1,5 @@
 const express = require('express')
 const router = express.Router()
 // const userController = require('../../controllers/user-controller')
-
+// const userController = require('../../../controllers/pages/user-controller')
 module.exports = router
